+++ CHANGELOG.md
@@ -3,6 +3,9 @@ This document describes changes between versions of rbxmk. The `imperative`
 branch is the latest unreleased version.
 
 ## imperative
+**Highlights:**
+- Add `rbxmk i` command to enter an interactive prompt.
+
 See a [comparison with the previous version][cmp-imperative] for a thorough list of changes.
 
 The [Documentation page][doc-imperative] provides a complete reference for this version of rbxmk.